refactor(builder): extract supervised build/optimize paths to AiModelBuilder.BuildPipeline.cs (#1427)
Pure mechanical partial-class split (no behaviour change): move BuildStreamingSupervisedAsync
+ BuildSupervisedInternalAsync (~2,323 LoC) out of the 9,486-LoC AiModelBuilder.cs into a new
partial file. Toward audit finding #12 (main file < 1,000 LoC). Core builds clean.
